Added cooldown between firing shots
Had some layer issues. Had to resort to ignoring the collision when the collision happens... Made a GameMaster to handle the target switching. Hasn't been implemented yet. TODO: Need to remove rb from bullet as it doesn't use it. Also need to work on the top spawner. // This project might end up being a solid prototype instead of a full released game. But we'll see. I like the state it's in.